requirements.txt
# PyTorch (CUDA 12.1 wheel – CUDA 12.4 드라이버에서 호환 가능)
# 설치 명령: pip install torch==2.3.0+cu121 torchvision==0.18.0+cu121 --index-url https://download.pytorch.org/whl/cu121
torch==2.3.0+cu121
torchvision==0.18.0+cu121

# OpenAI CLIP
git+https://github.com/openai/CLIP.git

# 이미지 처리 및 과학 연산
Pillow>=10.0.0
numpy==1.26.4
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0

# CLIP 종속성
ftfy>=6.0.0
regex>=2023.0.0

# GUI 라이브러리
tkinterdnd2>=0.3.0

# 기타 유틸리티
python-dateutil>=2.8.0